Added functions to retrieve the data from API endpoint "stories".
Changed the name of the mocked data state, the new state is now the main one.

Had to change ip adress of the API to be accessible from an extenal device, also made the change in Python. Will need to change this in prod.

Added a try catch, an alert and some console.log to be able to see the data I fetch.