Installed Lottie-React-Native package.
Added an animated loader before full loading of the map.

Added a button to ask for GPS permissions if they are not granted.

App now checks for permissions before render.

Added a modal to give informations to the user.

Changed the source of the image of StoryScreen.js from a source to an "uri" parameters, story images will be retrieved from API not in local.