Added one screen that will represent the account settings page for one user.
Started to work on the screens UserPageScreen.js and UserCollectionScreen.js but lot of works need to be done again. After connection, the user land on his collections of stories, one button will lead to his settings (password change, logout, account deletion). Logout is working.

Added some conditions on User collections to display either a message if null or some components if the user started to gather some stories in his account.

Stories request on HomeScreen.js is now filtered on the "validated" parameters in the query. It's closer to the "real-world" use case with the mandatory admin validation before publication. + request is now written with axios instead of fetch.

Added a title on connection section.