Created a navigator for the bottom tab left icon (only one screen for now). (changes in BottomTabNavigator.js and StackNavigator.js)
Alphabetically ordered styles in some screens.

Prepared the StoryScreen.js to include "liking" function. Heart Icon is now greyed out if user is not connected, an alert will warn him if he tries to like the story.

Small mistakes fixed.

Filled WriteStoryScreen.js with fields and states to prepare the work for the POST request of a story into database.

Started to work on image picker, installed the package and wrote some options.

Added a modal in a separate component to lighten WriteStoryScreen.js code. TODO : Do the same for HomeScreen.js modal.

Added a font file + some lines in AndroidManifest.xml