Added some permissions in AndroidManifest.xml.
New screens conditionnally rendered (user connected or not?) on touching left pen icon. One screen with the temporary icon of the app (added this file in assets) and a message encouraging to create an account before entering the page + an animation (added in assets too).

The other screen is the form to POST a new story, now working with some work needed to be done on error messages received from server + creation of 2 icons, one lauching camera, the other the gallery and not the dialog box as now. UseEffect used to refresh data when choosing a new image and set the states accordingly. Some work needed on the image render too. If there's no location data into picture data, geolocation will take over.